% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.transform.utils.R
\name{data.norms}
\alias{data.norms}
\title{\code{data.norms}}
\usage{
data.norms(X, type = NULL, center = F, scale = F, margin = 2)
}
\arguments{
\item{X}{a matrix for input}

\item{type}{the type of scaling to perform. Options: "ca" which is item divided by row sums, "hellinger" which is sqrt of "ca", "z" which is the same as scale(x), and "ss1" which is sum of squares 1. Also available is "scale" and requires use of center and scale parameters}

\item{center}{the intended center (see \code{\link{scale}})}

\item{scale}{the intended scale (see \code{\link{scale}})}

\item{margin}{which margin to perform this on (i.e., 1 for rows and 2 for columns)}
}
\value{
column or row normalized version of the matrix.
}
\description{
Compute normalization for rows or columns of a matrix.
}
