% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd.R
\docType{package}
\name{GSVD-package}
\alias{GSVD}
\alias{GSVD-package}
\title{GSVD}
\description{
The generalized singular value decomposition (GSVD) generalizes the standard SVD (see \code{\link{svd}}) procedure through addition of (optional) constraints applied to the rows and/or columns of a matrix.
}
\details{
While dedicated to the GSVD, this package also includes some nice features that are helpful for matrix analyses. For example there are tests to check if matrices are empty (\code{\link{is.empty.matrix}}) or identity (\code{\link{is.identity.matrix}}); also included are operations such as generalized inverse (\code{\link{matrix.generalized.inverse}}) and matrix exponents (\code{\link{matrix.exponent}} or \code{\link{\%^\%}}).
}
\examples{
see \\code{\\link{gsvd}}
}
\seealso{
\code{\link{gsvd}}, \code{\link{tolerance.svd}}, \code{\link{\%^\%}}
}
\author{
\strong{Maintainer}: Derek Beaton \email{exposition.software@gmail.com}

}
\keyword{component}
\keyword{decomposition}
\keyword{genearlized}
\keyword{matrix}
\keyword{multivariate}
\keyword{orthogonal}
\keyword{svd}
\keyword{variance}
