% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd.R
\name{gsvd}
\alias{gsvd}
\title{Generalized SVD}
\usage{
gsvd(DAT, LW, RW, k = 0, tol = .Machine$double.eps)
}
\arguments{
\item{DAT}{a data matrix to decompose}

\item{LW}{\bold{L}eft \bold{W}eights -- the constraints applied to the left side (rows) of the matrix and thus left singular vectors}

\item{RW}{\bold{R}ight \bold{W}eights -- the constraints applied to the right side (rows) of the matrix and thus right singular vectors}

\item{k}{total number of components to return though the full variance (based on nu and nv) will still be returned (see \code{Dv.orig})}

\item{tol}{default is .Machine$double.eps. A parameter with two roles: A tolerance level for (1) eliminating (tiny variance or negative or imaginary) components. (2) converting all values < tol to 0 in \item{u} and \item{v}}
}
\value{
A list with seven elements:
\item{u} Left singular vectors. A matrix whose columns contain the left singular vectors of x, present if nu > 0. Dimension c(n, nu) but also accounting for \code{tol}.
\item{v} Right singular vectors. A matrix whose columns contain the left singular vectors of x, present if nv > 0. Dimension c(p, nv) but also accounting for \code{tol}.
\item{p} Left generalized singular vectors. A vector containing the singular values of x of length min(n, p) but also accounting for \code{tol}.
\item{q} Right generalized singular vectors. A vector containing the singular values of x of length min(n, p) but also accounting for \code{tol}.
\item{d} a vector containing the singular values of x of length min(n, p) but also accounting for \code{tol}.
\item{d.orig} a vector containing the singular values of x of length min(n, p) but also accounting for \code{tol}.
\item{tau} a vector that contains the (original) explained variance per component.
}
\description{
\code{gsvd} takes in left (\code{LW}) and right (\code{RW}) constraints (usually diagonal matrices, but any positive semi-definite matrix is fine) that are applied to the data (\code{DAT})
  Left and right constraints are used for the orthogonality conditions.
}
\examples{
 ## an example with correspondence analysis.
 data(authors)
 author.data <- authors$ca$data
 Observed <- authors/sum(author.data)
 row.w <- rowSums(Observed)
   row.W <- diag(1/row.w)
 col.w <- colSums(Observed)
   col.W <- diag(1/col.w)
 Expected <- row.w \\\%o\\\% col.w
 Deviations <- Observed - Expected
 ca.res <- gsvd(Deviations,row.W,col.W)

 ## an example with canonical correlation analysis (though not all bells and whistles exist)
 data(two.table.wine)
 X <- scale(wine$objective)
 Y <- scale(wine$subjective)

 base.cca <- cancor(X,Y,F,F)

 cca.res <- gsvd(
     mgi(crossprod(X)) \\\%*\\\% t(X) \\\%*\\\% Y \%*\% mgi(crossprod(Y)),
     crossprod(X),
     crossprod(Y)
 )

 all.equal(base.cca$cor,cca.res$d)
 all.equal(base.cca$xcoef,cca.res$p)
 all.equal(base.cca$ycoef[,1:ncol(cca.res$q)],cca.res$q)

 cca.res$lx <- (X \%*\% cca.res$p)
 cca.res$ly <- (Y \%*\% cca.res$q)
 ### maximization:
 t(cca.res$lx) \%*\% cca.res$ly
 all.equal(diag(t(cca.res$lx) \%*\% cca.res$ly),cca.res$d)

}
\seealso{
\code{\link{tolerance.svd}} and \code{\link{svd}}
}
\author{
Derek Beaton
}
\keyword{diagonalization,}
\keyword{eigen}
\keyword{multivariate,}
