% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.diagonal.matrix.R
\name{is.diagonal.matrix}
\alias{is.diagonal.matrix}
\title{@export}
\usage{
is.diagonal.matrix(x, tol = .Machine$double.eps)
}
\description{
@title \code{is.diagonal.matrix}: test if a matrix is a diagonal matrix
}
\details{
@description \code{is.diagonal.matrix} takes a matrix and tests if it is a diagonal matrix.

@param x a matrix to test
 @param tol tolerance precision to eliminate all abs(x) values below \code{tol}. Default is .Machine$double.eps

@return
 boolean. TRUE if the matrix is a diagonal matrix, FALSE if the matrix is not.
}
