% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.empty.matrix.R
\name{is.empty.matrix}
\alias{is.empty.matrix}
\title{@export}
\usage{
is.empty.matrix(x, tol = .Machine$double.eps)
}
\description{
@title \code{is.empty.matrix}: test if a matrix contains all 0s
}
\details{
@description \code{is.empty.matrix} takes a matrix and tests if it contains all 0s

@param x a matrix to test
 @param tol tolerance precision to eliminate all abs(x) values below \code{tol}. Default is .Machine$double.eps

@return
 boolean. TRUE if the matrix contains all 0s, FALSE if the matrix does not.
}
