% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.identical.matrix.R
\name{is.identical.matrix}
\alias{is.identical.matrix}
\title{@export}
\usage{
is.identical.matrix(x, tol = .Machine$double.eps, round.digits = 12)
}
\description{
@title \code{is.identical.matrix}: test if a matrix contains all identical values
}
\details{
@description \code{is.identical.matrix} takes a matrix and tests if it contains all identical values

@param x a matrix to test
 @param tol tolerance precision to eliminate all abs(x) values below \code{tol}. Default is .Machine$double.eps
 @param round.digits number of decimal places to round to (via \code{\link{round}} function)

@return
 boolean. TRUE if the matrix contains all identical values, FALSE if the matrix does not.
}
