% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd.R
\name{gsvd}
\alias{gsvd}
\title{Generalized SVD}
\usage{
gsvd(DAT, LW, RW, k = 0, tol = .Machine$double.eps)
}
\arguments{
\item{DAT}{a data matrix to decompose}

\item{LW}{\bold{L}eft \bold{W}eights -- the constraints applied to the left side (rows) of the matrix and thus left singular vectors.}

\item{RW}{\bold{R}ight \bold{W}eights -- the constraints applied to the right side (rows) of the matrix and thus right singular vectors.}

\item{k}{total number of components to return though the full variance will still be returned (see \code{d.orig}). If 0, the full set of components are returned.}

\item{tol}{default is .Machine$double.eps. A parameter with two roles: A tolerance level for (1) eliminating (tiny variance or negative or imaginary) components and (2) converting all values < tol to 0 in \item{u} and \item{v}.}
}
\value{
A list with nine elements:
\item{d.orig}{A vector containing the singular values of DAT > \code{tol}.}
\item{tau}{A vector that contains the (original) explained variance per component (eigenvalues derived from \code{$d.orig}.}
\item{d}{A vector containing the singular values of x > \code{tol}. Length is \code{min(length(d.orig), k)}}
\item{u}{Left (rows) singular vectors. Dimensions are \code{nrow(DAT)} by k.}
\item{p}{Left (rows) generalized singular vectors. Dimensions are \code{nrow(DAT)} by k.}
\item{fi}{Left (rows) component scores. Dimensions are \code{nrow(DAT)} by k.}
\item{v}{Right (columns) singular vectors. Dimensions are \code{ncol(DAT)} by k.}
\item{q}{Right (columns) generalized singular vectors. Dimensions are \code{ncol(DAT)} by k.}
\item{fj}{Right (columns) component scores. Dimensions are \code{ncol(DAT)} by k.}
}
\description{
\code{gsvd} takes in left (\code{LW}) and right (\code{RW}) constraints (usually diagonal matrices, but any positive semi-definite matrix is fine) that are applied to the data (\code{DAT})
  Left and right constraints are used for the orthogonality conditions.
}
\examples{
 ## an example with correspondence analysis.
 data(authors)
 author.data <- authors$ca$data
 Observed <- author.data/sum(author.data)
 row.w <- rowSums(Observed)
   row.W <- diag(1/row.w)
 col.w <- colSums(Observed)
   col.W <- diag(1/col.w)
 Expected <- row.w \%o\% col.w
 Deviations <- Observed - Expected
 ca.res <- gsvd(Deviations,row.W,col.W)

 ## an example with canonical correlation analysis (though not all bells and whistles exist)
 data(two.table.wine)
 X <- scale(wine$objective)
 Y <- scale(wine$subjective)

 cca.res <- gsvd(
     matrix.generalized.inverse(crossprod(X)) \%*\% t(X) \%*\% Y \%*\% matrix.generalized.inverse(crossprod(Y)),
     crossprod(X),
     crossprod(Y)
 )

 cca.res$lx <- (X \%*\% cca.res$p)
 cca.res$ly <- (Y \%*\% cca.res$q)

 \dontrun{
     optimize.for <- t(cca.res$lx) \%*\% cca.res$ly
     all.equal(diag(optimize.for),cca.res$d)


     base.cca <- cancor(X,Y,F,F)

     sum(abs(base.cca$cor - cca.res$d)) < (.Machine$double.eps*100)
     base.cca$xcoef / cca.res$p
     base.cca$ycoef[,1:ncol(cca.res$q)] - cca.res$q
 }

}
\seealso{
\code{\link{tolerance.svd}} and \code{\link{svd}}
}
\author{
Derek Beaton
}
\keyword{diagonalization,}
\keyword{eigen}
\keyword{multivariate,}
